requirements.txt
torch
torchvision
diffusers
transformers
accelerate
wandb
matplotlib
pyyaml
torch-fidelity
scikit-image
tqdm
pillow
hydra-core
scikit-learn
random-fourier-features-pytorch